WRDL-112 Implement a method in the controller to complete the game

+++ src/main/java/ua/nure/wordle/dto/response/CabinetResponse.java
@@ -19,7 +19,7 @@ public class CabinetResponse {
     @JsonProperty("user_games")
     List<UserGameDTO> userGames;
     @NotNull
-    Integer wins;
+    Long wins;
     @NotNull
-    Integer losses;
+    Long losses;
 }
